TYPO FIX:
• The heuristic value and pathLength values stored in the Board obj MUST BE SIGNED. So their portable types are int32_t  (i.e. not uint32_t as I said before)

Why does this deserve a commit? Well, because I did them as straight INT instead of int32_t....!

+++ eight-puzzle/board_obj.cpp
@@ -162,7 +162,7 @@ std::ostream &Board::toStream(std::ostream &os) const {
   return os;
 }
 
-void Board::setHeuristic(const uint32_t &hVal) { heuristic = (int)hVal; }
+void Board::setHeuristic(const int32_t& hVal) { heuristic = hVal; }
 
 const uint32_t &Board::getHash(void) const { return hash; }
